mi: Add MPR callback facility
When a device sends a More Processing Required MI response, we may be
starting a long (blocking) delay in receiving the final command response
from the endpoint.

This change adds a small callback facility, to allow MI callers to
establish a "mpr" callback, invoked when we receive a MPR response to a
potentially long-running command.

This allows callers to set adequate timeouts, adjust the wait duration,
or even prevent waiting.

WIP: currently untested.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>